controller/requirements.txt
paho-mqtt
pygame
pynput